% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{bbox_ply}
\alias{bbox_ply}
\title{Create polygon from bounding box}
\usage{
bbox_ply(lon_min, lat_min, lon_max, lat_max)
}
\arguments{
\item{lon_min}{longitude, minimum}

\item{lat_min}{latitude, minimum}

\item{lon_max}{longitude, maximum}

\item{lat_max}{latitude, maximum}
}
\value{
Returns a spatial feature \link[sf]{sf} polygon.
}
\description{
Utility function to create polygon from bounding box
}
\examples{
# Florida Keys area
lon = -81.3; lat = 24.5; w = 10
bbox_ply(lon - w, lat - w, lon + w, lat + w)
}
\concept{read}
