% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seascape.R
\name{get_date_rng}
\alias{get_date_rng}
\title{Get date range of Seascape dataset}
\usage{
get_date_rng(info)
}
\arguments{
\item{info}{ERDDAP info object, as returned by \code{\link{get_seascape_info}})}
}
\value{
date range of min and max
}
\description{
Get date range of Seascape dataset
}
\examples{
s_info <- get_seascape_info()
get_date_rng(s_info)
}
