% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seascape.R
\name{get_seascape_data}
\alias{get_seascape_data}
\title{Get Seascape data}
\usage{
get_seascape_data(
  ctr_lon = -81.3,
  ctr_lat = 24.5,
  ctr_dd = 10,
  dataset = "global_monthly",
  var = "CLASS",
  date_beg = "2020-11-01",
  date_end = "2020-12-01"
)
}
\arguments{
\item{ctr_lon}{center longitude, in decimal degrees. Defaults to -81.3.}

\item{ctr_lat}{center longitude, in decimal degrees. Defaults to 24.5.}

\item{ctr_dd}{width around center, in decimal degrees. Defaults to 10.}

\item{dataset}{\verb{\\\{region\\\}_\\\{frequency\\\}} of dataset. Valid values (so far): "global_8day" or "global_monthly" (default).}

\item{var}{variable. One of "CLASS" (default) or "P" for probability.}

\item{date_beg}{date begin to fetch, as character (\code{"2003-01-15"}) or Date (\code{Date("2003-01-15")}).}

\item{date_end}{date end to fetch, as character (\code{"2020-11-15"}) or Date (\code{Date("2020-11-15")}).}
}
\value{
Raster \code{\link[raster]{raster}} layer if one date, \code{\link[raster]{stack}} if more
}
\description{
Get Seascape data
}
\examples{
r_monthly <- get_seascape_data(ctr_lon = -81.3, ctr_lat = 24.5, ctr_dd = 10, dataset = "global_monthly", var = "CLASS", date_beg = "2020-11-01", date_end = "2020-12-01")
r_monthly
r_8day <- get_seascape_data(ctr_lon = -81.3, ctr_lat = 24.5, ctr_dd = 10, dataset = "global_8day", var = "CLASS", date_beg = "2020-11-01", date_end = "2020-12-01")
r_8day
}
