% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{get_ss_dates}
\alias{get_ss_dates}
\title{Get date range of Seascape dataset}
\usage{
get_ss_dates(ss_info)
}
\arguments{
\item{ss_info}{ERDDAP info object, as returned by \code{\link{get_ss_info}})}
}
\value{
date range of min and max
}
\description{
Get date range of Seascape dataset
}
\examples{
ss_info <- get_ss_info()
get_ss_dates(ss_info)
}
\concept{read}
