% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{get_url_ply}
\alias{get_url_ply}
\title{Get polygon from link or NOAA Sanctuary code}
\usage{
get_url_ply(
  sanctuary = NULL,
  url = NULL,
  dir_ply = here::here("data_ss/ply"),
  verbose = F
)
}
\arguments{
\item{sanctuary}{NOAA Sanctuary code with which to form \code{url}. Sanctuary
codes can be found at
\href{https://sanctuaries.noaa.gov/library/imast_gis.html}{NOAA Sanctuaries
GIS}.}

\item{url}{Link from which to fetch a polygon. Originally designed to pull
from one of the zip links found at
\href{https://sanctuaries.noaa.gov/library/imast_gis.html}{NOAA Sanctuaries
GIS}.}

\item{dir_ply}{Directory to download locally into. This directory serves as
a cache to skip operations if already performed. If zip file exists, skip
downloading. If directory to unzip exists, skip unzipping.}

\item{verbose}{Verbose messages describing operations and choices, such as
skipping download or unzip and which shapefile used if more than one found.}
}
\value{
Returns a spatial feature \link[sf]{sf} polygon data frame.
}
\description{
Given a sanctuary code or link (ie URL), download the zip, unzip it, and read
first shapefile as a spatial feature. This function was originally designed
to pull from one of the zip links found at
\href{https://sanctuaries.noaa.gov/library/imast_gis.html}{NOAA Sanctuaries
GIS}.
}
\examples{
# mbnms: Monterey Bay National Marine Sanctuary
ply_mbnms <- get_url_ply(sanctuary = "mbnms")
ply_mbnms
plot(ply_mbnms[1])

# fknms: Florida Keys National Marine Sanctuary
ply_fknms <- get_url_ply(sanctuary = "fknms")
ply_fknms
plot(ply_fknms[1])
}
\concept{read}
