% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seascape.R
\name{map_seascape_raster}
\alias{map_seascape_raster}
\title{Map Seascape raster}
\usage{
map_seascape_raster(
  r,
  var = "CLASS",
  palette = "Spectral",
  basemap = providers$Esri.OceanBasemap,
  opacity = 0.8
)
}
\arguments{
\item{r}{raster layer}

\item{var}{variable. One of "CLASS" (default) or "P" for probability.}

\item{palette}{color palette. Default: "Spectral".}

\item{basemap}{basemap for leaflet. Default: \code{leaflet::providers$Esri.OceanBasemap}.}

\item{opacity}{transparency. Default: 0.8.}
}
\value{
Leaflet \code{\link[leaflet]{leaflet}} interactive map widget
}
\description{
Map Seascape raster
}
\examples{
r_monthly <- get_seascape_data(ctr_lon = -81.3, ctr_lat = 24.5, ctr_dd = 10, dataset = "global_monthly", var = "CLASS", date_beg = "2020-11-01", date_end = "2020-12-01")
map_seascape_raster(r_monthly)

r_8day <- get_seascape_data(ctr_lon = -81.3, ctr_lat = 24.5, ctr_dd = 10, dataset = "global_8day", var = "CLASS", date_beg = "2020-11-01", date_end = "2020-12-01")
r_8day
r_monthly

map_seascape_raster(raster::raster(r_8day, 1))
map_seascape_raster(raster::raster(r_8day, 2))
}
