% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seascape.R
\name{map_seascape_wms}
\alias{map_seascape_wms}
\title{Map Seascape using Web Map Server (WMS) image tiles}
\usage{
map_seascape_wms(
  date,
  ctr_lon = -81.3,
  ctr_lat = 24.5,
  ctr_dd = 10,
  var = "CLASS",
  basemap_style = "gbif-geyser"
)
}
\arguments{
\item{date}{date to map, eg \code{Date("2020-11-15")}}

\item{ctr_lon}{center longitude, in decimal degrees. Defaults to -81.3.}

\item{ctr_lat}{center longitude, in decimal degrees. Defaults to 24.5.}

\item{ctr_dd}{width around center, in decimal degrees. Defaults to 10.}

\item{var}{variable. One of "CLASS" (default) or "P" for probability.}

\item{basemap_style}{any style available from \href{https://tile.gbif.org/ui/}{GBIF background tiles}. Defaults to "gbif-geyser".}
}
\value{
Leaflet \code{\link[leaflet]{leaflet}} interactive map widget
}
\description{
Returns interactive map with Seascapes as image tiles (not data).
}
\details{
So far this uses the \href{https://tile.gbif.org/ui/}{GBIF background tiles}
since the ERDDAP WMS only displays in geographic project (EPSG = 4326).
}
\examples{
s_info <- get_seascape_info()

# map most recent Seascape
s_date_rng <- get_date_rng(s_info)
map_seascape_wms(s_date_rng[2], ctr_lon = -81.3, ctr_lat = 24.5, ctr_dd = 10)
}
