% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz.R
\name{map_ss_grd}
\alias{map_ss_grd}
\title{Map Seascape grid}
\usage{
map_ss_grd(
  grd,
  ply = NULL,
  grd_palette = "Spectral",
  ply_color = "blue",
  grd_opacity = 0.8,
  ply_opacity = 0,
  basemap = leaflet::providers$Esri.OceanBasemap
)
}
\arguments{
\item{grd}{grid}

\item{ply}{polygon}

\item{grd_palette}{color palette. Default: "Spectral".}

\item{ply_color}{polygon color}

\item{grd_opacity}{transparency of grid. Default: 0.8.}

\item{ply_opacity}{transparency of polygon fill color. Default: 0.}

\item{basemap}{basemap for leaflet. Default: \code{leaflet::providers$Esri.OceanBasemap}.}
}
\value{
Leaflet \code{\link[leaflet]{leaflet}} interactive map widget
}
\description{
Map Seascape grid with optional polygon as interactive map.
}
\examples{
ply  <- get_url_ply("mbnms")
ss_i <- get_ss_info()
grds <- get_ss_grds(ss_i, ply, date_beg = "2020-01-01")

# get first grid
grd  <- raster::raster(grds, 1)

map_ss_grd(grd)

}
\concept{viz}
