% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz.R
\name{map_ss_wms}
\alias{map_ss_wms}
\title{Map Seascape image tiles from web map server (wms)}
\usage{
map_ss_wms(
  ss_info,
  ply = NULL,
  date = max(get_ss_dates(ss_info)),
  ss_var = "CLASS",
  ply_color = "blue",
  ply_opacity = 0.2,
  basemap_style = "gbif-geyser"
)
}
\arguments{
\item{ss_info}{SeaScape ERDDAP info object, as returned by \code{\link{get_ss_info}})}

\item{ply}{optional polygon as spatial feature \code{\link[sf]{sf}}, as returned by
\code{\link{get_url_ply}} or \code{\link{bbox_to_ply}}, to use for initial extent of map and optionally plot (see \code{ply_color}).}

\item{date}{date of seascape variable to map. Defaults to most recent date available from \code{ss_info}, eg \code{Date("2020-11-15")}.}

\item{ss_var}{SeaScape variable. One of \code{"CLASS"} (default) or \code{"P"} for probability.}

\item{ply_color}{polygon color. To not plot polygon and just use for initial extent of map set \code{ply_color = NA}. Default: \code{"blue"}.}

\item{ply_opacity}{transparency of polygon Default: \code{0.2}.}

\item{basemap_style}{any style available from \href{https://tile.gbif.org/ui/}{GBIF background tiles}. Defaults to \code{"gbif-geyser"}.}
}
\value{
Leaflet \code{\link[leaflet]{leaflet}} interactive map widget
}
\description{
Returns interactive map with Seascapes as image tiles (not data) from web map server (wms).
}
\details{
So far this uses the \href{https://tile.gbif.org/ui/}{GBIF background tiles}
since the ERDDAP WMS only displays in geographic project (EPSG = 4326).
}
\examples{
ss_info <- get_ss_info()

# map most recent Seascape
s_dates <- get_date_rng(ss_info)
map_ss_wms(s_dates[2], ctr_lon = -81.3, ctr_lat = 24.5, ctr_dd = 10)
}
\concept{viz}
