% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nms}
\alias{nms}
\title{National Marine Sanctuaries list}
\format{
A list with keys and values indicating:
\describe{
\item{\code{code}}{Short lower-case Sanctuary code used by \code{get_url_ply()}.}
\item{\code{name}}{Full name, capital case.}
}
}
\source{
\url{https://sanctuaries.noaa.gov/library/imast_gis.html}
}
\usage{
nms
}
\description{
National Marine Sanctuaries list of short code and full name.
}
\concept{data}
\keyword{datasets}
