% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seascape.R
\name{plot_seascape_ts}
\alias{plot_seascape_ts}
\title{Plot Seascape time series}
\usage{
plot_seascape_ts(r, show_legend = "follow")
}
\arguments{
\item{r}{raster stack with more than one date, as returned by
\code{\link{get_seascape_data}}}

\item{show_legend}{When to display the legend. Specify "follow" (default) to
have the legend show as overlay to the chart which follows the mouse.
Specify "always" to always show the legend. Specify "onmouseover" to only
display it when a user mouses over the chart. The "auto" option results in
"always" when more than one series is plotted and "onmouseover" when only a
single series is plotted.}
}
\value{
\code{\link[dygraphs]{dygraph}} interactive plot
}
\description{
Plot Seascape time series
}
\examples{
r_mo_2019 <- get_seascape_data(
  ctr_lon = -81.3, ctr_lat = 24.5, ctr_dd = 10,
  dataset = "global_monthly", var = "CLASS",
  date_beg = "2019-01-01", date_end = "2020-01-01")

plot_seascape_ts(r_mo_2019)

}
