% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz.R
\name{plot_ss_class_var_svg}
\alias{plot_ss_class_var_svg}
\title{Plot Seascape Class variable as SVG}
\usage{
plot_ss_class_var_svg(var, val, g_svg, web_svg, redo = F, ht_ratio = 0.3, ...)
}
\arguments{
\item{var}{variable to plot, for getting distribution of all Class averages from \code{\link{ss_gl_classes}}}

\item{val}{average value to plot as a vertical line}

\item{g_svg}{local file path to output ggplot as scalable vector graphic (*.svg)}

\item{web_svg}{web prefix with which to return the path in HTML as an image}

\item{ht_ratio}{ratio of height to width; default = \code{0.3}}

\item{...}{other parameters to pass onto \code{\link{plot_ss_class_var}}}
}
\value{

}
\description{
This function wraps around \code{\link{plot_ss_class_var}} to write the ggplot as a vector
format file (*.svg) and return a path to the file that could be used for the
web.
}
\concept{viz}
