% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz.R
\name{plot_ss_ts}
\alias{plot_ss_ts}
\title{Plot Seascape time series from table}
\usage{
plot_ss_ts(tbl, sum_var = "pct_cells", fillAlpha = 0.8, show_legend = "follow")
}
\arguments{
\item{tbl}{table in same format as output by \code{sum_ss_grds_to_ts}.}

\item{sum_var}{column name of summary variable to display; one of percent
(\code{"pct_cells"}; default) or number of cells (\code{"n_cells"}).}

\item{show_legend}{When to display the legend. Specify "follow" (default) to
have the legend show as overlay to the chart which follows the mouse.
Specify "always" to always show the legend. Specify "onmouseover" to only
display it when a user mouses over the chart. The "auto" option results in
"always" when more than one series is plotted and "onmouseover" when only a
single series is plotted.}
}
\value{
\code{\link[dygraphs]{dygraph}} interactive plot
}
\description{
Plor Seascape as an interactive stacked and filled time series plot.
}
\examples{

# prep input table of time series data
ply  <- get_url_ply("mbnms")
ss_i <- get_ss_info("global_8day")
grds <- get_ss_grds(ss_i, ply, date_beg = "2020-09-01")
tbl  <- sum_ss_grds_to_ts(grds)

plot_ss_ts(tbl)

}
\concept{viz}
