% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ss_vars}
\alias{ss_vars}
\title{SeaScape Variables}
\format{
A table of average values per class:
\describe{
\item{\code{var}}{original variable name in \code{\link{ss_gl_classes}}}
\item{\code{v}}{short variable name, useful for file naming}
\item{\code{description}}{text description}}
}
\usage{
ss_vars
}
\description{
Basic information about variables used to inform SeaScape Classes
}
\concept{data}
\keyword{datasets}
