% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{get_ss_grds}
\alias{get_ss_grds}
\title{Get Seascape grids within polygon for date range}
\usage{
get_ss_grds(
  ss_info,
  ply,
  ss_var = "CLASS",
  date_beg = min(get_ss_dates(ss_info)),
  date_end = max(get_ss_dates(ss_info)),
  dir_tif = NULL
)
}
\arguments{
\item{ss_info}{SeaScape ERDDAP info object, as returned by \code{\link{get_ss_info}})}

\item{ply}{polygon as spatial feature \code{\link[sf]{sf}}, as returned by
\code{\link{get_url_ply}} or \code{\link{bbox_to_ply}}}

\item{ss_var}{SeaScape variable. One of "CLASS" (default) or "P" for probability.}

\item{date_beg}{date begin to fetch, as character (\code{"2003-01-15"}) or Date
(\code{Date("2003-01-15")}). Defaults to first date available from \code{ss_info}.}

\item{date_end}{date end to fetch, as character (\code{"2020-11-15"}) or Date
(\code{Date("2020-11-15")}). Defaults to latest date available from \code{ss_info}.}
}
\value{
Raster \code{\link[raster]{raster}} layer if one date,
\code{\link[raster]{stack}} if more
}
\description{
Given a polygon and date range, fetch Seascape data and return a raster layer
for a single date or raster stack if multiple dates found for given date range.
}
\examples{
ply  <- get_url_ply("mbnms")
ss_i <- get_ss_info()
grds <- get_ss_grds(ss_i, ply, date_beg = "2020-01-01")
grds

}
\concept{read}
