% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz.R
\name{plot_ss_ts}
\alias{plot_ss_ts}
\title{Plot Seascape time series from table}
\usage{
plot_ss_ts(tbl, colors = NULL, fillAlpha = 0.8, show_legend = "always")
}
\arguments{
\item{tbl}{table in same format as output by \code{sum_ss_grds_to_ts} containing
columns: \code{date} (date), \code{cellvalue} (double), \code{n_cells} (double).}

\item{colors}{Character vector of colors for the data series. See
\code{dygraphs::dyOptions()}. The default (\code{NULL}) automatically generates colors
from the "Spectral" color ramp (see \code{RColorBrewer::brewer.pal()}).}

\item{show_legend}{When to display the legend. See \code{dygraphs::dyLegend()}.}
}
\value{
\code{\link[dygraphs]{dygraph}} interactive plot
}
\description{
Plot Seascape as an interactive stacked and filled time series plot.
}
\examples{

# prep input table of time series data
ply  <- get_url_ply("mbnms")
ss_i <- get_ss_info("global_8day")
grds <- get_ss_grds(ss_i, ply, date_beg = "2020-09-01")
tbl  <- sum_ss_grds_to_ts(grds)

plot_ss_ts(tbl)

}
\concept{viz}
