% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ss_gl_classes}
\alias{ss_gl_classes}
\title{SeaScape Global Classes}
\format{
A table of average values per class:
\describe{
\item{\code{CLASS}}{unique integer identifier of the Class}
\item{\code{NAME}}{descriptive name}
\item{\verb{SST  (°C)}}{sea surface temperature}
\item{\code{SSS (psu)}}{sea surface salinity}
\item{\code{ADT (m)}}{TBD}
\item{\verb{ICE (\%)}}{percent ice, if present}
\item{\code{CDOM (m^-1)}}{colored dissolved organic matter}
\item{\verb{CHLA (mg m^-3)}}{Chlorophyll a}
\item{\verb{NFLH (W m^-2 um^-1 sr^-1)}}{TBD}
\item{\code{NFLH:CHL}}{TBD}
\item{\code{LATITUDE}}{latitude}
\item{\verb{DOMINANT HEMISPHERE}}{TBD}
\item{\verb{DOMINANT SEASON}}{TBD}
}
}
\source{
\url{https://sanctuaries.noaa.gov/library/imast_gis.html}
}
\usage{
ss_gl_classes
}
\description{
Average properties of Seascape Global Classes.
}
\concept{data}
\keyword{datasets}
