% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze.R
\name{sum_ss_grds_to_ts}
\alias{sum_ss_grds_to_ts}
\title{Summarize Seascape grids into time series}
\usage{
sum_ss_grds_to_ts(grds, ts_csv = NULL)
}
\arguments{
\item{grds}{raster stack with more than one date, as returned by
\code{\link{get_ss_grds}}}
}
\value{
\code{tibble} of data
}
\description{
Summarize Seascape grids into a table of: \code{date}, number of cells (\verb{ncells_*}) and percent (\verb{pct_*}) of variable
(eg \code{ss_var = "CLASS"} or \code{"P"}).
}
\details{
This function is particularly helpful in between using \code{get_ss_grds()} and \code{plot_ss_ts()}.

Note that the \code{ncells_NA} gets subtracted from the overall minimum number of NAs to account for masked area.
}
\examples{
ply  <- get_url_ply("mbnms")
ss_i <- get_ss_info()
grds <- get_ss_grds(ss_i, ply, date_beg = "2020-01-01")
tbl  <- sum_ss_grds_to_ts(grds)
tbl

}
\concept{analyze}
